+++ lib/rubocop/cop/style/word_array.rb
@@ -8,7 +8,31 @@ module RuboCop
       #
       # Alternatively, it can check for uses of the %w() syntax, in projects
       # which do not want to include that syntax.
+      #
+      # Configuration option: MinSize
+      # If set, arrays with fewer elements than this value will not trigger the
+      # cop. For example, a `MinSize of `3` will not enforce a style on an array
+      # of 2 or fewer elements.
+      #
+      # @example
+      #   EnforcedStyle: percent (default)
+      #
+      #   # good
+      #   %w[foo bar baz]
+      #
+      #   # bad
+      #   ['foo', 'bar', 'baz']
+      #
+      # @example
+      #   EnforcedStyle: brackets
+      #
+      #   # good
+      #   ['foo', 'bar', 'baz']
+      #
+      #   # bad
+      #   %w[foo bar baz]
       class WordArray < Cop
+        include ArrayMinSize
         include ArraySyntax
         include ConfigurableEnforcedStyle
         include PercentLiteral
@@ -23,9 +47,9 @@ module RuboCop
 
         def on_array(node)
           if bracketed_array_of?(:str, node)
-            check_bracketed(node)
+            check_bracketed_array(node)
           elsif node.percent_literal?(:string)
-            check_percent(node)
+            check_percent_array(node)
           end
         end
 
@@ -39,18 +63,17 @@ module RuboCop
 
         private
 
-        def check_bracketed(node)
+        def check_bracketed_array(node)
           return if complex_content?(node.values) ||
-                    comments_in_array?(node)
-          style_detected(:brackets, node.values.size)
+                    comments_in_array?(node) ||
+                    below_array_length?(node)
 
-          return unless style == :percent && node.values.size >= min_size
-
-          add_offense(node, :expression, PERCENT_MSG)
+          array_style_detected(:brackets, node.values.size)
+          add_offense(node, :expression, PERCENT_MSG) if style == :percent
         end
 
-        def check_percent(node)
-          style_detected(:percent, node.values.size)
+        def check_percent_array(node)
+          array_style_detected(:percent, node.values.size)
           add_offense(node, :expression, ARRAY_MSG) if style == :brackets
         end
 
@@ -74,10 +97,6 @@ module RuboCop
           end
         end
 
-        def min_size
-          cop_config['MinSize']
-        end
-
         def word_regex
           Regexp.new(cop_config['WordRegex'])
         end
@@ -88,45 +107,6 @@ module RuboCop
           lambda do |corrector|
             corrector.replace(node.source_range, "[#{words.join(', ')}]")
           end
-        end
-
-        def style_detected(style, ary_size)
-          cfg = config_to_allow_offenses
-          return if cfg['Enabled'] == false
-
-          largest_brackets = largest_brackets_size(style, ary_size)
-          smallest_percent = smallest_percent_size(style, ary_size)
-
-          if cfg['EnforcedStyle'] == style.to_s
-            # do nothing
-          elsif cfg['EnforcedStyle'].nil?
-            cfg['EnforcedStyle'] = style.to_s
-          elsif smallest_percent <= largest_brackets
-            self.config_to_allow_offenses = { 'Enabled' => false }
-          else
-            cfg['EnforcedStyle'] = 'percent'
-            cfg['MinSize'] = largest_brackets + 1
-          end
-        end
-
-        def largest_brackets_size(style, ary_size)
-          self.class.largest_brackets ||= -Float::INFINITY
-
-          if style == :brackets && ary_size > self.class.largest_brackets
-            self.class.largest_brackets = ary_size
-          end
-
-          self.class.largest_brackets
-        end
-
-        def smallest_percent_size(style, ary_size)
-          @smallest_percent ||= Float::INFINITY
-
-          if style == :percent && ary_size < @smallest_percent
-            @smallest_percent = ary_size
-          end
-
-          @smallest_percent
         end
       end
     end